VR: Optimize Header and update CLI interface (breaking)
This is a breaking change for client and cluster init that:

* adds `Header.head` as a u128,
* adds `Header.version` as a u8,
* resizes `Header.cluster` from u128 to u32,
* accepts `Header.cluster` now being zero (no longer a UUID),
* resizes `Header.view` from u64 to u32,
* resizes `Header.replica` from u16 to u8, and
* adds a client table fix.

The motivation behind this is to:

* add space in the header to support end-to-end integrity
  checking in the client to match replies back to requests,
  and prevent future issues with protocol versioning, and

* make the cluster CLI more memorable and easier to use:

  --cluster=0 --addresses=3001,3002,3003 --replica=0

+++ src/main.zig
@@ -45,7 +45,7 @@ pub fn main() !void {
     var replica = try Replica.init(
         arena,
         args.cluster,
-        @intCast(u16, args.configuration.len),
+        @intCast(u8, args.configuration.len),
         args.replica,
         &journal,
         &message_bus,
